backend/requirements.txt
# Core
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-dotenv>=1.0.1
pydantic>=2.10.3
pydantic-settings>=2.7.0

# Scheduler
APScheduler==3.10.4

# Google Drive
google-api-python-client==2.155.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# OpenAI (cheapest: gpt-4o-mini)
openai==1.58.1

requests==2.32.3

# Telegram Bot
python-telegram-bot>=21.0

# Database
# SQLite is built-in to Python — no extra package needed

# Utils
httpx==0.28.1
python-multipart==0.0.20
